# Backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TallerMecanico | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5290/"
    }
  ],
  "paths": {
    "/api/v1/especialidad": {
      "get": {
        "tags": [
          "Especialidad"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Especialidade"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Especialidad"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EspecialidadDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/especialidad/{id}": {
      "get": {
        "tags": [
          "Especialidad"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Especialidad"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EspecialidadDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Especialidad"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/mecanicos": {
      "get": {
        "tags": [
          "Mecanicos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mecanico"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Mecanicos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MecanicoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/mecanicos/{id}": {
      "get": {
        "tags": [
          "Mecanicos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Mecanicos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MecanicoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Mecanicos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/modelo": {
      "get": {
        "tags": [
          "Modelo"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Modelo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/modelo/{id}": {
      "get": {
        "tags": [
          "Modelo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Modelo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Modelo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cliente": {
        "type": "object",
        "properties": {
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "correo": {
            "type": [
              "null",
              "string"
            ]
          },
          "direccion": {
            "type": [
              "null",
              "string"
            ]
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          },
          "vehiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehiculo"
            }
          }
        }
      },
      "ClienteDTO": {
        "required": [
          "nombre",
          "apellido",
          "correo",
          "telefono",
          "direccion"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "apellido": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          }
        }
      },
      "DetalleRepuesto": {
        "type": "object",
        "properties": {
          "detalleRepuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "repuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "precioUnitario": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          },
          "repuesto": {
            "$ref": "#/components/schemas/Repuesto"
          }
        }
      },
      "DetalleServicio": {
        "type": "object",
        "properties": {
          "detalleServicioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "servicioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "precioUnitario": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          },
          "servicio": {
            "$ref": "#/components/schemas/Servicio"
          }
        }
      },
      "EspecialidadDTO": {
        "required": [
          "nombre",
          "descripcion"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "descripcion": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "Especialidade": {
        "type": "object",
        "properties": {
          "especialidadId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "mecanicos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mecanico"
            }
          }
        }
      },
...
      "Mecanico": {
        "type": "object",
        "properties": {
          "mecanicoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "especialidadId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "especialidad": {
            "$ref": "#/components/schemas/Especialidade"
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          }
        }
      },
      "MecanicoDTO": {
        "required": [
          "nombre",
          "apellido",
          "telefono",
          "especialidadId"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "apellido": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "telefono": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "especialidadId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ModeloDTO": {
        "required": [
          "nombre",
          "marcaId"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "marcaId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "OrdenesTrabajo": {
        "type": "object",
        "properties": {
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fechaIngreso": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "observaciones": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "vehiculoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mecanicoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "detalleRepuestos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleRepuesto"
            }
          },
          "detalleServicios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleServicio"
            }
          },
          "mecanico": {
            "$ref": "#/components/schemas/Mecanico"
          },
          "pago": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pago"
              }
            ]
          },
          "vehiculo": {
            "$ref": "#/components/schemas/Vehiculo"
          }
        }
      },
...
      "Pago": {
        "type": "object",
        "properties": {
          "pagoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "montoTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "metodoPago": {
            "type": "string"
          },
          "fechaPago": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          }
        }
      },
...
      "Repuesto": {
        "type": "object",
        "properties": {
          "repuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "precio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "stock": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detalleRepuestos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleRepuesto"
            }
          }
        }
      },
...
      "Vehiculo": {
        "type": "object",
        "properties": {
          "vehiculoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "patente": {
            "type": "string"
          },
          "anio": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "modeloId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "modelo": {
            "$ref": "#/components/schemas/Modelo"
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          }
        }
      },
...
    }
  },
...
}